rename html directories
