+++ tests/website/test_website.py
@@ -1,7 +1,7 @@
 import pytest
 from pathlib import Path
 
-from website_analysis.html.website import Website
+from website_analysis.website.website import Website
 
 @pytest.fixture
 def imagesSite():
